TISCv2.2 Revised `pcr` instruction, included new `peek` instruction
v2.1 had introduced the `gtr` instruction, which was too similar to the
`pcr` instruction. v2.2 Revised `pcr` instruction to push to the stack,
which reduces the required number of bytes for calling routines.

Updated the `terminal_test.tac` program to print a welcome string upon
startup, print a terminal character when ready to accept input. Added a
new command `clear` to clear the TTY.